test: split backendDevice schema tests into per-module test files
Move embed schema tests to embed-schemas.test.ts and completion-stream
schema tests to completion-stream-schemas.test.ts, matching the SDK
convention of naming test files after the module they cover.
